breaking things while prepping serdes and print_2d

+++ cpp/src/solvent_util/math.hpp
@@ -1,9 +1,9 @@
 #ifndef HPP_SOLVENT_UTIL__PROB
 #define HPP_SOLVENT_UTIL__PROB
 
-[[nodiscard, gnu::const]] constexpr unsigned long n_choose_r(unsigned int n, unsigned int r) noexcept {
+[[nodiscard, gnu::const]] constexpr unsigned long long n_choose_r(unsigned int n, unsigned int r) noexcept {
 	if (r > n - r) r = n - r;
-	unsigned long ans = 1;
+	unsigned long long ans = 1;
 	for (unsigned int i = 1; i <= r; ++i) {
 		ans *= n - r + i;
 		ans /= i;
